webserver/requirements.txt
tensorflow
keras
Pillow
redis
python-multipart
